+++ CHANGELOG.md
@@ -2,6 +2,7 @@ medialibrary CHANGELOG
 ======================
 
 ## CURRENT
+ - [medialibrary-63](https://github.com/cjcodeproj/medialibrary/issues/63) Create object code to handle "art" proper noun
  - [medialibrary-62](https://github.com/cjcodeproj/medialibrary/issues/62) First PyPi test
  - [medialibrary-56](https://github.com/cjcodeproj/medialibrary/issues/56) File loader ignored Blu-Ray 3d movies
  - [medialibrary-45](https://github.com/cjcodeproj/medialibrary/issues/45) XML elements without text may cause an error when stripped